ml/dl/examples/pid_project_extract/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alabaster==0.7.13
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.7.1
arrow==1.2.3
astor==0.8.1
astroid==2.15.5
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
autoflake==1.7.8
Automat==22.10.0
Babel==2.12.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bandit==1.7.5
bce-python-sdk==0.8.87
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==22.12.0
bleach==6.0.0
blinker==1.6.2
build==0.10.0
bump2version==1.0.1
bumpversion==0.6.0
CacheControl==0.12.14
cachetools==5.3.1
certifi==2023.7.22
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.2.0
ci-info==0.3.0
cleo==2.0.1
click==8.1.6
codecov==2.1.13
colorama==0.4.6
commonmark==0.9.1
conan==2.0.9
constantly==15.1.0
contourpy==1.1.0
cookiecutter==1.7.3
coverage==7.2.7
crashtest==0.4.1
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.0
darglint==1.8.1
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
dnspython==2.4.2
docstring-parser==0.15
docutils==0.18.1
dulwich==0.21.5
email-validator==2.0.0.post2
eradicate==2.3.0
et-xmlfile==1.1.0
etelemetry==0.3.0
exceptiongroup==1.1.1
fastapi==0.100.1
fastapi-template==5.0.3
fasteners==0.18
filelock==3.12.0
fire==0.5.0
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.5.0
flake8-bugbear==22.12.6
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.4.0
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
Flask==2.3.2
flask-babel==3.1.0
fonttools==4.41.1
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
humanize==4.6.0
hyperlink==21.0.0
identify==2.5.26
idna==3.4
imageio==2.31.1
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==6.6.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
interval==1.0.0
iopath==0.1.10
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.13.1
kiwisolver==1.4.4
layoutparser@ https://paddleocr.bj.bcebos.com/whl/layoutparser-0.0.0-py3-none-any.whl#sha256=0e3026045cd88f791012cc8b52bbd9c1c3b14f161ec3c2855e52aff3a14167b8
lazy-object-proxy==1.9.0
lazy_loader==0.3
lmdb==1.4.1
lockfile==0.12.2
looseversion==1.3.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.6.1
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nibabel==5.1.0
nodeenv==1.8.0
numpy==1.25.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.5.1
paddlepaddle-gpu==2.5.1
pandas==2.0.3
passlib==1.7.4
patch-ng==1.17.4
pathlib==1.0.1
pathspec==0.11.2
pbr==5.11.1
pdf2docx==0.5.6
pdf2image==1.16.3
pefile==2023.2.7
pep8-naming==0.13.2
pexpect==4.8.0
piccolo==0.117.0
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
portalocker==2.7.0
poyo==0.5.0
pre-commit==3.3.3
premailer==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.2
prov==2.0.0
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycodestyle==2.8.0
pycryptodome==3.18.0
pydantic==1.10.12
pydantic_core==2.4.0
pydocstyle==6.3.0
pydot==1.4.2
pyenchant==3.2.2
pyflakes==2.4.0
Pygments==2.16.1
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pylint==2.17.4
PyMuPDF==1.20.2
pyparsing==3.1.1
pyproject_api==1.5.1
pyproject_hooks==1.0.0
pyte==0.8.1
pytest==7.3.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-env==0.8.2
pytest-runner==6.0.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-pptx==0.6.21
python-slugify==8.0.1
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
qualname==0.1.0
rapidfuzz==2.15.1
rarfile==4.0
rdflib==7.0.0
readme-renderer==37.3
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
restructuredtext-lint==1.4.0
rfc3986==1.5.0
rich==13.5.2
rpds-py==0.9.2
scikit-image==0.21.0
scipy==1.11.1
service-identity==21.1.0
shapely==2.0.1
shellingham==1.5.0.post1
simple-term-menu==1.6.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==6.2.1
sphinx-autoapi==2.1.0
sphinx-rtd-theme==1.2.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-images==0.9.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starlette==0.27.0
stevedore==5.1.0
targ==0.3.8
termcolor==1.1.0
text-unidecode==1.3
thefuck==3.32
tifffile==2023.7.18
tomli==2.0.1
tomlkit==0.11.8
tox==4.6.0
tox-gh-actions==3.1.1
tqdm==4.65.0
traits==6.3.2
trove-classifiers==2023.8.7
twine==4.0.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typer==0.7.0
tzdata==2023.3
ujson==5.8.0
Unidecode==1.3.6
urllib3==2.0.4
uvicorn==0.22.0
virtualenv==20.23.0
visualdl==2.5.3
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
wemake-python-styleguide==0.17.0
Werkzeug==2.3.6
win-unicode-console==0.5
wrapt==1.15.0
XlsxWriter==3.1.2
yapf==0.29.0
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
